switchcase

+++ Arthematic Operation.c
@@ -4,14 +4,14 @@ multiplication, division and modulo) */
 #include<stdio.h>
 main()
 {
-	int a,b;
+	float a,b;
 	printf ("Enter value of a : ");
-	scanf("%d",&a);
+	scanf("%f",&a);
 	printf("Enter Value b : ");
-	scanf("%d",&b);
-	printf("%d + %d = %d",a,b,a+b);
-	printf("\n%d - %d = %d",a,b,a-b);
-	printf("\n%d * %d = %d",a,b,a*b);	
-	printf("\n%d / %d = %d",a,b,a/b);
-	printf("\n Mod = %d",a%b);
+	scanf("%f",&b);
+	printf("%f + %f = %f",a,b,a+b);
+	printf("\n%d - %d = %f",a,b,a-b);
+	printf("\n%d * %d = %f",a,b,a*b);	
+	printf("\n%d / %d = %f",a,b,a/b);
+	printf("\n Mod = %f",a%b);
 }
